Add more to release automation workflow
This commit updates more details on the release automation workflow. It
finishes the first draft of bumping the meta repo. The next step here is
to add file locking around git repo access so that we can ensure that a
git repo is only being modified by one thing at a time. After that this
workflow actually needs to be tested.