Add interprocess locking for git and api operations
Since the bot will be listening to webhooks from multiple repositories
eventually it will be good to set up interprocess locking for a world
where this is deployed for real. If multiple webhook requests come into
a real server at once it will likely fork to handle the concurrent
requests. Adding locks around per repo operations will make sure that we
don't trample on a parallel operations for something that is expecting
an unchanging state (like updating a repo, pushing a branch, and
creating a PR from that branch).