Fix issues from local testing of startup
Now that there is a sandbox environment setup to configure things
manually for testing. Actually running things have found some basic
issues during initial setup of the daemon. This commit addresses those
so now it can start a bot daemon and handle requests to '/' on the
server (which will trigger local repo creation).