Handle existing repositories on daemon startup
The initialization of the daemon goes over the config file and creates a
new Repo object for each configured repository. However, in the case of
a daemon restart this would fail because the creation of a new local
clone would fail because it already existed. This commit handles this
edge case by not trying to create a repo if the directory already
exists. This enables restarting the daemon safely without needing to
delete the local working directory.