Fix oversight on tag creation check
The event handler for create events was incorrectly checking for a tag
by only checking if the ref_type field existed instead of checking
whether it was a tag or not. This ended up meaning that the bot was
treating every create event as a new release. This commit fixes the
issue so that tags create events are the only one treated as releases.