Add logging and make minor version sticky
This commit adds two small details to the release processing. The first
was a minor issue if a minor meta repo version bump is pending and
patch release comes in the minor version should take priority. However
prior to this patch it just checked that the new version was different
and used the new version if it was. The second is to add additional
logging around meta repo version numbers to debug when a PR is generated
with an incorrect version.