Update docs conf version on meta bump PR too
This commit updates the local file modifications done during the
generation of the bump meta repository PR to also include bumping the
docs version string in docs/conf.py. This is needed when we increase the
version in the setup.py to make sure the docs version reflects the
current meta-package release.