Add support for a repo local config file
This commit adds support for setting a local repo config file for each
repo tracked by the bot. This will let you set per repo configuration
that will be change frequently in the repo itself. The starting 2 fields
are 'users', which will be used by the eventual PR workflow to set
approver permissions and notifications, and 'categories' which is used
to set mappings from PR tags to the Changelog category that is used for
release (which was previously in the global config). This new model
should enable us to dynamically configure these 2 aspects of the
workflow so we can adjust their usage as the project evolves. Moving
forward if there are any other use cases for the bot this will give us a
place to add per repo settings.