Drop python 3.5 support
The qiskit-bot is not being written as a general purpose tool, it's
designed to be run specifically for upstream qiskit development and
releases. There will only ever be one deployment of it which to start is
running python 3.7. Since the syntax used for regular expression is not
compatible with python 3.5 there is no reason to even try to support it,
so this commit removes support for running in python 3.5 completely.