Log stdout and stderr on git exceptions
Right now when a local git command fails for some reason we log the
exception, but there is no detail on the output just the command that
failed and the error code from git. This makes debugging a failure next
to impossible. To improve this situation this commit adds the stdout and
stderr from the exception log message to the error message so that we
can debug this issues when they're encountered.