Add config option to set log level and log format
When running qiskit-bot it's logging isn't super useful by default. It
only shows errors and warnings by default and doesn't have any
formatting. This makes it difficult to debug issues that come up when
running it. This commit adds config options to enable setting the log
level and output format so that we can set these when running the api in
production. In addition it sets sane defaults for running productions so
normally you won't have to change the config.