Make meta repo required
While we're setting things up there seems to be an issue with the
default value for the meta_repo config option. To test the production
bot it needs to talk to the sandbox meta repository, not Qiskit/qiskit,
while we iron out issues. This commit makes the field required without a
default to make setting the meta_repo always explicit.