Add calls to update AUTHORS and bibtex on meta bump
This commit adds the necessary calls during the bump meta function to
update the AUTHORS list and update the bibtex file. These 2 files should
be update on every release of the meta package to ensure we keep the
lists up to date. This commit adds calls to the 2 scripts used to update
these files in the meta repo as part of the normal workflow for
generating the PR to bump the meta repo.

For this to work properly we'll need Qiskit/qiskit#570 to merge.