Remove duplicate microseconds field from default log format
This commit updates the default log format used by the qiskit-bot to
remove a duplicate microseconds field. The microseconds are included in
the normal time field (and we really don't need that resolution anyway).
To cleanup the format and make the logs a little smaller this removes
the duplicate field.