Explicitly use 'python' instead of sys.executable
When running a wsgi app under uwsgi sys.executable becomes the path to
the uwsgi binary. This is because uwsgi links against libpython for the
purpose of running a wsgi app and adding it's python extensions to apps
that use it. However, this ends up breaking any calls that expect a
python binary as the executable returned by sys.executable. This commit
updates the calls to generate_authors.py and generate_bibtex.py to use
python explicitly instead of sys.executable to avoid hitting this issue.