Launch bump meta in a subprocess
With the addition of running generate_authors and generate_bibtex the
bump_meta function is taking too long. It goes past the standard timeout
for the http response from flask to the webhook. Even though everything
succeeds we're returning errors to github for the webhook event. This
commit attempts to address this by switching to launching the bump_meta
in a separate process. This should make the blocking operations for
sending the response code very small and allow a quick turnaround for
sending a successful return code to github.