Add tool to generate changelog manually
This commit adds a tool to manually generate a changelog outside of a
release. It works at any point on the master branch and will return the
changelog from any tag (or sha1) specified on the command line. At the
same time a small fix was made to the changelog generation to fix an
issue if there was no PR associated with the commit.