Do changelog generation in a separate process
The changelog generation was previously happening synchronously. This
can however take a while because it can involve hundreds of github api
calls to look up each PR that's been merged for a particular release.
The time this takes can exceed the HTTP timeout on the webhook, which
will result in an error on the github side. This isn't critical because
it doesn't stop the bot from running or cause any errors.