Add missing changelog entries to manual script
This commit adds a list of commits not included in the changelog to
the manual changelog generation script. That script is used for
verifying the list is correct and without seeing what's potentially
missing a tag it's hard to know if we got everything. By showing this in
the manual script it becomes easier to identify if we need to add or
remove tags prior to a release.