Fix fall through logic for missing changelog entries
This commit fixes an issue in the logic for detecing whether a commit is
missing from the changelog. Previously we were incorrectly adding it if
the label was not present on the commit message. This was wrong because
the commit would be added to the missing list for each non-changelog
label on it. This is not the correct behavior and could result in a
commit being in the missing list multiple times (if there were multiple
non-changelog labels) or being added to the list even if there were
valid changelog tags on it. This fixes the logic by setting a boolean
flag when a pr has a changelog tag, and only adding it to the missing
list after we loop over the labels if the boolean flag is still not set.