Handle invalid PR numbers in changelog generation
In the upcoming qiskit-terra release a large chunk of qiskit-aqua's
history was migrated into qiskit-terra. These commit messages use
qiskit-aqua's commit summary line in the form of
"Qiskit/qiskit-aqua#1234" which breaks PR lookup because the regex
match returns that full string which raises a Value error when later we
try to convert the match string into an int. This commit avoids that
error by catching the value error and just skipping that PR lookup,
since it isn't relevant for the changelog anyway.