Add a configuration option for optional packages
With the recent release of the qiskit-nature, qiskit-optimization,
qiskit-machine-learning, and qiskit-finance packages we now have qiskit
packages in the metapackage as optional extras. We still want to track
these packages in the metapackage so that users can install everything
together with known good working versions, but since they're not
installed by default when they release we only want to bump the
metapackage patch version and never bump the metaversion. This commit
adds a new per repo config option 'optional_package' which when set will
mean when the bump_meta function is triggered for that repo means the
metapackage bump_meta PR will only ever increment the patch version (if
necessary).