Add support for a configurable default branch
There has been a recent push to rename the default branch for git
repositories to be something more inclusive than 'master'. Qiskit would
like to make this transition and rename the default branch to 'main'.
However, qiskit-bot has been blocking this transition because it
unconditionally was using 'master' for generating pull requests,
updating local checkouts, etc. To address this and enable us to
transition away from using 'master' for our default branch this commit
adds a new optional per repo configuration option 'default_branch' which
is used to set the default branch for a repository. If not set it
defaults to using master so there is no breaking change for the deployed
bot and we can update the configuration as we rename the branches.
Similarly for the metapackage repository a new optional configuration
option, 'meta_repo_default_branch', is added that behaves the same way
but for the meta_repo.