Add option for changelog none to changelog generator
This commit adds a new option to the default changelog categories dict
(and by extension a custom changelog categories config file that has a
value as `None` for the category) for "Changelog: None". PRs labelled
with this entry will be excluded from both the generated changelog
categories and the missing list if the standalone changelog generator
script is run.

Fixes #7