Add option to set default branch on generate_changelog
This commit adds a new argument to the generate_changelog script,
--default-branch/-b, which is used to specify the default branch for
running the changelog generation on. Previously this was hardcoded to
only run on master which doesn't work to the new default branch we use
on the majority of projects, main. The new option is set to use main as
the default (since it's what most qiskit projects use currently) and for
projects that have not migrated to main yet the flag can be used to
specify master or any other default branch.