Migrate to github actions
Travis is no longer a viable service for running CI on qiskit-bot,
mainly because their tight per user quotas prevent CI from reliably
running for every contributor. This commit moves the CI configuration
from travis to run in github actions which doesn't have this issue (and
is generally more reliable and faster). This way we'll have a reliable
CI solution for testing proposed changes to qiskit-bot.