Change behavior for optional_package repo option
This commit changes the behavior of the optional_package repo config
option. This option originally was used to assume an optional package
had a pinned version as an optional extra in the setup.py for the qiskit
metapackage and would bump that to the new release and only ever bump
the patch version (as it wasn't actually a tracked requirement of the
metapackage). However this wasn't desireable because qiskit shouldn't
have a dependency on downstream projects like this. We provide the
optional extras with downstream qiskit projects as a user convenience
but incompatibilities between qiskit and the downstream qisit packages
should be resolved in the downstream project (or handled as fixes to
qiskit). Accordingly this changes the behavior to never bump the
metapackage in case of an optional_package release. The optional
extras in the metapackage should managed manually like other
non-requirements package metadata moving forward.

As an aside this option can now also be used by other qiskit projects
that want the other automation features (like changelog generation on
release) which aren't in the metapackage at all (optional or required).
For example, we could now use this for retworkx releases if we needed
the automation features offered by qiskit-bot by using the
optional_package config setting.