Add support for configurable notification comments
This commit adds a new feature for enabling per repo configurable
notification comments to qiskit-bot. When a repository has a per repo
configuration file (ie qiskit_bot.yaml in the repo root) with the
notifications section set this can be used to provide a list of
usernames that will get mentioned in a generated comment automatically
based on the files modified in the PR. This lets each project have a
local configuration of who gets notified of changes to each submodule in
the repository.

Fixes #5