Fix unittests after updating default notification prelude
In the previous commit we updated the default notification prelude to
fix a typo, but that commit neglected to update the tests accordingly.
This caused all the tests to error because the expected string didn't
match what was actually used. This commit fixes that oversight and fixes
the test failures.