Add support for configurable notification comments (#18)
* Add support for configurable notification comments

This commit adds a new feature for enabling per repo configurable
notification comments to qiskit-bot. When a repository has a per repo
configuration file (ie qiskit_bot.yaml in the repo root) with the
notifications section set this can be used to provide a list of
usernames that will get mentioned in a generated comment automatically
based on the files modified in the PR. This lets each project have a
local configuration of who gets notified of changes to each submodule in
the repository.

Fixes #5

* Fix Python 3.7 unittest compatibility

* Fix attribute access on PyGithub File objects

The filename attribute on PyGithub's File class is filename, not
file_name. This typo caused things to crash when trying to get the list
of changed files. This commit fixes this oversight.

* Finish default prelude text

* Fix lint and default prelude text

* Add option for always notifying even without matching files

This commit adds a new option to the local per repo config file for
always notifying on PR opening even if there are no matching paths
to mention users in the message. If this flag is set true it will always
at least print the configured prelude.

* Add README section on in-repo configuration

This commit adds a new section to the README documenting how to
configure qiskit-bot's behavior on a per repo level by creating a
`qiskit_bot.yaml` configuration file. It shows each option and how they
influence the behavior of the bot.

* Update qiskit_bot/notifications.py

Co-authored-by: Manoel Marques <manoelmrqs@gmail.com>

* Fix unittests after updating default notification prelude

In the previous commit we updated the default notification prelude to
fix a typo, but that commit neglected to update the tests accordingly.
This caused all the tests to error because the expected string didn't
match what was actually used. This commit fixes that oversight and fixes
the test failures.

* Fix typos found in code review

This commit fixes typos in various documentation, text blocks,
and docstrings which were found during code review.

Co-authored-by: Jake Lishman <jake@binhbar.com>

* Update copyright year in qiskit_bot/notifications.py

Co-authored-by: Jake Lishman <jake@binhbar.com>

* Update tests for recent prelude changes

* Remove fixed 80 character line wrap from default notification message

* Update README.md

Co-authored-by: Jake Lishman <jake@binhbar.com>

* Update default message in README

Co-authored-by: Manoel Marques <manoelmrqs@gmail.com>
Co-authored-by: Jake Lishman <jake@binhbar.com>