Add support for handling pre-releases
This commit adds support to the release automation for dealing with
pre-releases. A pre-release tag should trigger branch creation (if
enabled) but not trigger a metapackage bump PR. This commit adds the
packaging dependency to determine if a PEP440 comatible pre-release
is tagged and if so handle the tag as a pre-release. The pre-release
event will create a prerelease github release page with the changelog
from the appropriate previous full release and create a branch from the
tag if the repo is configured to branch on release.

Fixes #21