Fix lint with flake8 5.x.x
A new flake8 release, 5.x.x, changed some of the detection behavior
around line length rules that is now triggering on a multiline string
where it wasn't before. This is blocking CI and lint from running
passing locally. This commit fixes the formatting so as to avoid
triggering the rule but still work as expected. This should fix the CI
and local lint job failures and work with both new and old versions of
flake8.