Use privilege to determine user association (#29)
The webhook data that the bot gets fed with automatically only includes
"public" data.  Many members of organisations are private, so their
association appears to be "none" or "contributor" instead of "member"
or "owner", and the bot would overzealously tag things as being from an
external member of the community.  Instead, we can use the bot's API
authorisation slightly earlier in the process to get the priviledged
data.