Fix compatibility with tox 4
With the release of tox 4.0.0 tox underwent a large internal refactor
and made a number of backwards incompatible behavior changes. This commit
updates the tox.ini so that it will work in tox >= 4.0.0. The primary
change is to remove the filtering of the cached python bytecode files
which shouldn't have been needed anymore anyway (it was a legacy thing I
copied from another repo that was still supporting python 2.7 at the
time).