Fix metapackage doc config handling after conf.py updates (#40)
In Qiskit/qiskit-metapackage#1689 the formatting of the docs/conf.py
file in the metapackage was updated to use black. This had the side
effect of breaking the "parsing" the bot does to update the version
string in the docs/conf.py file as part of the metapackage bumping
procedure. This commit updates the regex used to match the conf.py
file. Ideally we would change the handling here to be a bit more robust
to changes (although to be fair this is the first breaking change in the
formatting of that file since qiskit-bot's introduction in Aug 2019) but
there are only a handful of metapackage release left so at this point
it's not worth spending too much effort maintaining that part of the
release process.