Cap flask version
Our use of `before_first_request` decorator to lazily initialize the
repositories and configuration of the webapp was deprecated and removed
in the 2.3.x release series. We'll have to reorganize the api module to
conform to more modern flask conventions to do this initial setup. In
the meantime this commit just caps our flask requirement to avoid 2.3.x.

This issue was never caught in the deployed service because we were
deploying the bot using Python 3.7 which effectively was capping us
flask<2.3. Trying to update the deployment to use Python 3.11 instead
the service never starts because 2.3 is being installed.