Handle major versions and non-rc prereleases (#43)
* Handle major versions and non-rc prereleases

This commit adds support to the qiskit-bot release automation to handle
major version releases and non-rc pre-releases. The major version
handling is primarily for building the log command, as we need to get
the log from 'n.0.0..n-1.m.0' where m is the most recent non-prerelease.
To handle non-rc prereleases the only change is we don't create a new
branch for the pre-release. Previously all pre-releases would trigger
branch creation (given the appropriate configuration variable). But the
desired behavior is to actually only create a stable branch from a
release candidate.

* Normalize rc detection

* Update comments

* Fix tests to account for PEP440